refactore - Create a navMobile component to add a new animation
