+++ src/const/index.js
@@ -1,0 +1,16 @@
+// import { ToggleLang } from '../toggleLang/ToggleLang';
+import { BiHomeAlt2 } from "react-icons/bi";
+import { FiSearch } from "react-icons/fi";
+
+export const routes = [
+  {
+    title: "mainNav.home",
+    href: "/",
+    Icon: BiHomeAlt2,
+  },
+  {
+    title: "mainNav.projects",
+    href: "/projects",
+    Icon: FiSearch,
+  }
+];
